Fix OAuth2 implementation for FastMCP compatibility and test suite
- Remove **kwargs from MCP tool functions (FastMCP constraint)
- Simplify OAuth2 decorator to synchronous operation
- Fix JWT PEM conversion to return bytes instead of string
- Replace complex OAuth2 tests with basic functionality tests
- Ensure all 50 tests pass including OAuth2 authentication
- Maintain backward compatibility with OAuth2 disabled by default
