+++ Pipfile
@@ -5,7 +5,7 @@ name = "pypi"
 
 [packages]
 pandas = "*"
-m3 = {file = ".", editable = true}
+m3 = {extras = ["dev"], file = ".", editable = true}
 
 [dev-packages]
